sdks/apigw-manager/pyproject.toml
[tool.poetry]
name = "apigw-manager"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.6.1"
setuptools = ">=21.0.0"
urllib3 = ">=1.25.3"
future = "0.18.2"
pyyaml = "^5.4.1"
pyjwt = ">=1.6.4"
python-dateutil = "2.8.1,<3.0.0"
bkapi-bk-apigateway = "^1.0.3"
bkapi-client-core = "^1.0.8"

[tool.poetry.extras]
cryptography = ["cryptography>=3.1.1"]
django=["django>=1.11.1"]

[tool.poetry.dev-dependencies]
m2r = "^0.2.1"
pytest = "^6.2.2"
pytest-cov = ">=2.8.1"
black = "^20.8b1"
isort = "^5.8.0"
autoflake = "^1.4"
pre-commit = "^2.11.1"
mypy = "^0.812"
flake8 = "^3.9.0"
Faker = "^6.6.3"
pytest-mock = "^3.5.1"
tox = "^3.23.0"
ipython = "7.16.1"
jedi = "0.17.2"
pytest-django = "^4.1.0"
python_dateutil = ">=2.5.3"
twine = "^3.4.1"

[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
default = true
secondary = false

[[tool.poetry.source]]
name = "tencent-blueking"
url = "https://mirrors.tencent.com/tencent_pypi/simple/"
secondary = true

[tool.black]
line-length = 119
skip-string-normalization = 'true'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

